BAH-4821 | Refactor. Move groupExtensionsByPoint To Extensions Utils
